LPS-106881 Split the test, move testHasSubscriptionsReturnsFalseWhenNoSubscriptionsInDB and testHasSubscriptionsReturnsTrueWhenSubscriptionsInDB to integration test because SubscriptionSender.hasSubscribers will call com.liferay.subscription.service.SubscriptionLocalServiceUtil and the static block of SubscriptionLocalServiceUtil depends on integration test environment
